third_party/blink/web_tests/external/wpt/css/css-contain/contain-style-dynamic-001-expected.txt

This is a testharness.js-based test.
[FAIL] switching contain from none to style
  assert_equals: increment-counter expected true but got false
[FAIL] switching contain from style to none
  assert_equals: increment-counter expected false but got true
Harness: the test ran to completion.
